Adding application messaging.
While widgets are the most common target for events, there is good reason to sometimes target the application
as a whole.  This patch adds support for such events using the /-prefixed notation.  (This convention is safe
for use since it represents an extremely uncommon widget ID, and one which data could not be POSTed to.)
Such events are NOT subject to replay, and are handled on the client-side by firing corresponding events on
the Dashing application class.

As a proof of concept (and to fufill a feature request), this patch also introduces a POST /reload endpoint,
which when provided with a valid authkey (and optionally a dashboard name) fires the appropriate 'reload'
event on all of the loaded dashboards.  The 'reload' event handler then reloads the dashboard, unless a
different dashboard was specified.