Fix: calling set with same value twice breaks update
When you push a value that is equal to the prior value when truncated to an integer, the call to `clearInterval()` never happens so further updates are continually overwritten by the interval that was never cleared. This allows for values that are equal to pass through and clear the previous interval.